@@ -1,14 +1,13 @@
-import {Response} from '@angular/http'
+import {HttpErrorResponse} from '@angular/common/http'
 import {Observable} from 'rxjs/Observable'
 import 'rxjs/add/observable/throw'
 
 export class ErrorHandler {
-  static handleError(error: Response | any){
+  static handleError(error: HttpErrorResponse | any){
     let errorMessage: string
-    if (error instanceof Response){
-      const body = error.json() || ''
-      const err = body.error || JSON.stringify(body)
-      errorMessage = `${error.url}: ${error.status} - ${error.statusText || ''} ${err}`
+    if (error instanceof HttpErrorResponse){
+      const body = error.error
+      errorMessage = `${error.url}: ${error.status} - ${error.statusText || ''} ${body}`
     }else{
       errorMessage = error.message ? error.message : error.toString()
     }
