@@ -5,6 +5,7 @@ import { RestaurantDetailComponent } from './restaurant-detail/restaurant-detail
 import { MenuComponent } from './restaurant-detail/menu/menu.component';
 import { ReviewsComponent } from './restaurant-detail/reviews/reviews.component';
 import { OrderSummaryComponent } from './order-summary/order-summary.component';
+import { NotFoundComponent } from "./not-found/not-found.component";
 
 export const ROUTES: Routes = [
   { path: '', component: HomeComponent },
@@ -19,5 +20,7 @@ export const ROUTES: Routes = [
   },
   { path: 'order', loadChildren: './order/order.module#OrderModule' },
   {path: 'order-summary', component: OrderSummaryComponent},
-  {path: 'about', loadChildren: './about/about.module#AboutModule'}
+  {path: 'about', loadChildren: './about/about.module#AboutModule'},
+  {path: '**', component: NotFoundComponent}
+                       
 ]
