@@ -1,7 +1,8 @@
+import { Observable } from 'rxjs/Observable';
 import { Injectable } from '@angular/core'
 
-import { Http, Headers, RequestOptions } from '@angular/http'
-import { Observable } from 'rxjs/Observable'
+import { HttpClient} from '@angular/common/http'
+ 
 import 'rxjs/add/operator/map'
 
 import { ShoppingCartService } from '../restaurant-detail/shopping-cart/shopping-cart.service'
@@ -9,12 +10,12 @@ import { CartItem } from '../restaurant-detail/shopping-cart/cart-item.model'
 
 
 import { MEAT_API } from '../app.api'
-import { Order, OrderItem } from './order.model';
+import { Order} from './order.model';
 
 @Injectable()
 export class OrderService {
 
-  constructor(private cartService: ShoppingCartService, private http: Http) { }
+  constructor(private cartService: ShoppingCartService, private http: HttpClient) { }
 
   itemsValue(): number {
     return this.cartService.total()
@@ -41,12 +42,7 @@ export class OrderService {
   }
 
   checkOrder(order: Order): Observable<string> {
-    const headers = new Headers()
-    headers.append('Content-Type', 'application/json')
-    return this.http.post(`${MEAT_API}/orders`,
-      JSON.stringify(order),
-      new RequestOptions({ headers: headers }))
-      .map(response => response.json())
+    return this.http.post<Order>(`${MEAT_API}/orders`,order)      
       .map(order => order.id)
   }
 }
