@@ -1,5 +1,5 @@
+import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Http } from '@angular/http';
 import { Restaurant } from './restaurant/restaurant.model';
 import { MEAT_API } from './../app.api';
 import { Observable } from 'rxjs/Observable';
@@ -10,31 +10,28 @@ import { MenuItem } from '../restaurant-detail/menu-item/menu-item.model';
 
 @Injectable()
 export class RestaurantsService {
-    constructor(private http: Http){
+    constructor(private http: HttpClient) {
 
     }
-    
-    restaurants():Observable<Restaurant[]>{
-     return this.http.get(`${MEAT_API}/restaurants`)
-        .map(response => response.json())
-        .catch(ErrorHandler.handleError);
+
+    restaurants(search?: string): Observable<Restaurant[]> {
+        let parms: HttpParams = undefined
+        if(search){
+            parms = new HttpParams().set('q', search);
+        }
+
+        return this.http.get<Restaurant[]>(`${MEAT_API}/restaurants`, { params: parms })
     }
 
-    restaurantById(id: string): Observable<Restaurant>{
-      return this.http.get(`${MEAT_API}/restaurants/${id}`)
-        .map(response => response.json())
-        .catch(ErrorHandler.handleError)
+    restaurantById(id: string): Observable<Restaurant> {
+        return this.http.get<Restaurant>(`${MEAT_API}/restaurants/${id}`)
     }
 
-    reviewsOfRestaurant(id: string): Observable<any>{
+    reviewsOfRestaurant(id: string): Observable<any> {
         return this.http.get(`${MEAT_API}/restaurants/${id}/reviews`)
-          .map(response => response.json())
-          .catch(ErrorHandler.handleError)
-      }
+    }
 
-      menuOfRestaurant(id: string): Observable<MenuItem[]>{
-        return this.http.get(`${MEAT_API}/restaurants/${id}/menu`)
-          .map(response => response.json())
-          .catch(ErrorHandler.handleError)
-      }
+    menuOfRestaurant(id: string): Observable<MenuItem[]> {
+        return this.http.get<MenuItem[]>(`${MEAT_API}/restaurants/${id}/menu`)
+    }
 }
